==> notebooks/3-construct-toy-example.ipynb <==
import pandas as pd
import numpy as np
import csv

# Define MIMIC demo data path
data_path = $your_mimic_datapath$

# Load tables
input_events = pd.read_csv(data_path+'INPUTEVENTS_MV.csv', header=0)
input_events.head()

procedures = pd.read_csv(data_path+'PROCEDURES_ICD.csv', header=0)
procedures.head()

# Group by 'subject_id', and sort by 'starttime' and then 'itemid'
drug_events_only = input_events.groupby(by='subject_id').apply(lambda x: x.sort_values('starttime'))['itemid'].reset_index(level=[1], drop=True)

drug_events_by_patient = drug_events_only.groupby(by='subject_id').apply(list)
drug_events_by_patient[:5]

# Group by subject_id and sort by admittime (first) and seq_num (second)
procedure_codes = procedures.groupby(by='subject_id').apply(lambda x: x.sort_values(['hadm_id', 'seq_num']))['icd9_code'].reset_index(level=[1], drop=True)

procedures_by_patient = procedure_codes.groupby(by='subject_id', axis=0).apply(list)
procedures_by_patient[:5]

# Merge drug events and procedures
drug_events_procedures_merged = pd.merge(drug_events_by_patient, procedures_by_patient, how='inner', on='subject_id')

drug_events_procedures_merged.head()

drug_events_procedures_merged.shape

# Load patients table
patients = pd.read_csv(data_path+'PATIENTS.csv', header=0)
patients.head()

# expire_flag: 1 indicates death in the hospital, and 0 indicates survival to hospital discharge.
sum(patients['expire_flag'] == 1) # no patients survived in the demo data

# Make a sample with replacement, to construct the toy example
pos_data = drug_events_procedures_merged.sample(n=100, random_state=3, replace=True)
neg_data = drug_events_procedures_merged.sample(n=100, random_state=1, replace=True)

pos_data.head()

neg_data.head()

pos_data2 = pd.DataFrame()

pos_data2['drug_events'] = pos_data['itemid'].apply(lambda x: ' '.join(str(i) for i in x))
pos_data2['procedure_codes'] = pos_data['icd9_code'].apply(lambda x: ' '.join(str(i) for i in x))

# Export to folder './toy_example/' 
pos_data2.to_csv(path_or_buf='./toy_example/train_pos.txt', index=False, header=False, sep=' ', quoting = csv.QUOTE_NONE, escapechar = ' ')
pos_data2.to_csv(path_or_buf='./toy_example/validation_pos.txt', index=False, header=False, sep=' ', quoting = csv.QUOTE_NONE, escapechar = ' ')

neg_data2 = pd.DataFrame()

neg_data2['drug_events'] = neg_data['itemid'].apply(lambda x: ' '.join(str(i) for i in x))
neg_data2['procedure_codes'] = neg_data['icd9_code'].apply(lambda x: ' '.join(str(i) for i in x))

neg_data2.to_csv(path_or_buf='./toy_example/train_neg.txt', index=False, header=False, sep=' ', quoting = csv.QUOTE_NONE, escapechar = ' ')
neg_data2.to_csv(path_or_buf='./toy_example/validation_neg.txt', index=False, header=False, sep=' ', quoting = csv.QUOTE_NONE, escapechar = ' ')
